chore: move security related classes to Security package
