+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/configuration/AuditorAwareImpl.java
@@ -1,6 +1,6 @@
 package eu.bbmri.eric.csit.service.negotiator.configuration;
 
-import eu.bbmri.eric.csit.service.negotiator.configuration.auth.NegotiatorUserDetails;
+import eu.bbmri.eric.csit.service.negotiator.configuration.security.auth.NegotiatorUserDetails;
 import eu.bbmri.eric.csit.service.negotiator.database.model.Person;
 import java.util.Optional;
 import org.springframework.data.domain.AuditorAware;
