+++ src/main/java/eu/bbmri/eric/csit/service/negotiator/configuration/security/auth/JwtUserDetails.java
@@ -1,4 +1,4 @@
-package eu.bbmri.eric.csit.service.negotiator.configuration.auth;
+package eu.bbmri.eric.csit.service.negotiator.configuration.security.auth;
 
 import eu.bbmri.eric.csit.service.negotiator.database.model.Person;
 import java.util.Collection;
